Rename GitSettingsForm to GitAuthenticationForm and update component
