+++ tests/unit/components/form/GitAuthenticationForm.spec.js
@@ -1,23 +1,22 @@
 import { installQuasarPlugin } from '@quasar/quasar-app-extension-testing-unit-jest';
 import { shallowMount } from '@vue/test-utils';
-import GitSettingsForm from 'src/components/form/GitSettingsForm.vue';
+import GitAuthenticationForm from 'src/components/form/GitAuthenticationForm.vue';
 import {
   getProjectById,
   saveProject,
-  gitAddRemote,
 } from 'src/composables/Project';
 import { Notify } from 'quasar';
 import GitEvent from 'src/composables/events/GitEvent';
 
-jest.mock('src/composables/events/GitEvent', () => ({
-  AddRemoteEvent: {
-    next: jest.fn(),
-  },
-}));
-
 installQuasarPlugin({
   plugins: [Notify],
 });
+
+jest.mock('src/composables/events/GitEvent', () => ({
+  AuthenticationEvent: {
+    next: jest.fn(),
+  },
+}));
 
 jest.mock('vue-i18n', () => ({
   useI18n: () => ({
@@ -40,22 +39,16 @@ jest.mock('src/composables/Project', () => ({
     };
   }),
   saveProject: jest.fn(),
-  gitAddRemote: jest.fn((project) => {
-    if (project.id === 'error') {
-      return Promise.reject({ name: 'HttpError' });
-    }
-    return Promise.resolve();
-  }),
 }));
 
-describe('Test component: GitSettingsForm', () => {
+describe('Test component: GitAuthenticationForm', () => {
   let wrapper;
   const emit = jest.fn();
 
-  GitEvent.AddRemoteEvent.next.mockImplementation(() => emit());
+  GitEvent.AuthenticationEvent.next.mockImplementation(() => emit());
 
   beforeEach(() => {
-    wrapper = shallowMount(GitSettingsForm, {
+    wrapper = shallowMount(GitAuthenticationForm, {
       props: {
         projectName: 'test',
       },
@@ -69,24 +62,9 @@ describe('Test component: GitSettingsForm', () => {
       });
     });
 
-    describe('Test variable: repository', () => {
-      it('should be undefined without any git configuration in project', () => {
-        wrapper = shallowMount(GitSettingsForm, {
-          props: {
-            projectName: 'noGit',
-          },
-        });
-        expect(wrapper.vm.repository).toBeUndefined();
-      });
-
-      it('should not be null with git configuration in project', () => {
-        expect(wrapper.vm.repository).toEqual('testRepository');
-      });
-    });
-
     describe('Test variable: token', () => {
       it('should be undefined without any git configuration in project', () => {
-        wrapper = shallowMount(GitSettingsForm, {
+        wrapper = shallowMount(GitAuthenticationForm, {
           props: {
             projectName: 'noGit',
           },
@@ -101,7 +79,7 @@ describe('Test component: GitSettingsForm', () => {
 
     describe('Test variable: username', () => {
       it('should be undefined without any git configuration in project', () => {
-        wrapper = shallowMount(GitSettingsForm, {
+        wrapper = shallowMount(GitAuthenticationForm, {
           props: {
             projectName: 'noGit',
           },
@@ -116,7 +94,7 @@ describe('Test component: GitSettingsForm', () => {
   });
 
   describe('Test function: onSubmit', () => {
-    it('should emit an event on success', async () => {
+    it('should emit an event and send a positive notification', async () => {
       wrapper.vm.repository = 'https://test/test.git';
       wrapper.vm.token = 'testToken';
       wrapper.vm.username = 'TestUsername';
@@ -127,31 +105,10 @@ describe('Test component: GitSettingsForm', () => {
 
       expect(getProjectById).toBeCalled();
       expect(saveProject).toBeCalled();
-      expect(gitAddRemote).toBeCalled();
       expect(wrapper.vm.submitting).toEqual(false);
       expect(wrapper.emitted()['project-git:save']).toBeTruthy();
       expect(emit).toBeCalled();
       expect(Notify.create).toHaveBeenCalledWith(expect.objectContaining({ type: 'positive' }));
-    });
-
-    it('should emit a notification on error', async () => {
-      wrapper = shallowMount(GitSettingsForm, {
-        props: {
-          projectName: 'error',
-        },
-      });
-
-      Notify.create = jest.fn();
-
-      wrapper.vm.repository = 'https://test/test.git';
-      wrapper.vm.token = 'testToken';
-      wrapper.vm.username = 'TestUsername';
-      await wrapper.vm.onSubmit();
-
-      expect(getProjectById).toBeCalled();
-      expect(saveProject).toBeCalled();
-      expect(gitAddRemote).toBeCalled();
-      expect(Notify.create).toHaveBeenCalledWith(expect.objectContaining({ type: 'warning' }));
     });
   });
 });
